pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "projeto-redes-neurais"
version = "0.1.0"
description = "Pipeline inicial para estudo e comparacao de embeddings com CelebA-Spoof."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pillow>=10.3",
  "scikit-image>=0.24",
  "scikit-learn>=1.5",
  "torch>=2.3",
  "torchvision>=0.18",
  "matplotlib>=3.9",
  "seaborn>=0.13",
  "tqdm>=4.66",
  "umap-learn>=0.5.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
